+++ app/src/api/types/incoming/sharedFolder.ts
@@ -1,0 +1,16 @@
+// SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
+//
+// SPDX-License-Identifier: EUPL-1.2
+import { NamespacedIncoming } from '@opentalk/common';
+
+export interface UpdateSharedFolder {
+  message: 'updated';
+  read: string;
+  readWrite?: string;
+}
+
+export type Message = UpdateSharedFolder;
+
+export type SharedFolder = NamespacedIncoming<Message, 'shared_folder'>;
+
+export default SharedFolder;