+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Get rid of login popup and remove redux-persist (!1030)
 - Fix error handling on update displayname in dashboard (!1026)
+- Fix unhandled shared folder update message (!1010)
 - Remove settings if provided by a service provider (!1000)
 - Add missing names to toolbar buttons (!988)
 - Change tariff to plan in english translation (!1004)